perf: don't walk the input tree if all files are explicitly defined
If we have a big input tree and only need to pick explicitly defined files,
it should not be necessary to walk the entire input tree. Instead, take the
 property and use that as the complete list of input files.

I didn't include any tests since only internal implementation has changed
and it's not clear how could I mock out the walkSync function to verify that
it wasn't called.