Ensure adding additional trees does not affect include/exclude results.
In a7ed4fa82a6492728cd48cbe3fb0b48c79d4c3ae we added support for passing
additional trees to the constructor (to ensure that they can be ready to
read in subclasses), but we didn't ensure that all internal usage of
`this.input` was limited to the first tree passed. This didn't matter
much for the simple symlink srcDir <-> destDir situation, but as soon as
you use custom `include`/`exclude` globs (and we therefore have to walk
all the files) we would trigger errors.

This fixes that oversight by ensuring that all usage of `this.input` is
scoped to `this.input.at(0)`, and adds a previously failing test.